Disallow vectors in source
They need to be handled by category theoretic `list` in zlang. Will use the
Scheme object of a vector to represent closures internally. Only problem is that
there is a read syntax for them, so clean the file input early on to ensure
there are no vectors directly from the source.

The only other disallowed Scheme type is a port, which has no read syntax so
there is no potential for source code to include it.